backend/requirements.txt
django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djoser==2.0.1
gunicorn==20.1.0
pillow==9.2.0
psycopg2-binary==2.8.6
